Optimize HMAC calculation in interleaved AES-CBC+HMAC-SHA256
Currently for every EVP Update call, we calculate the final HMAC tag
which involves 2 hash calculations. This is unnecessary as the
subsequent Update calls will overwrite the tag. The final tag can be
calculated only once in the end. This patch optimizes this HMAC
calculation. For Update calls, we store the internal values of hash1
and return. Only when EVP final call is made, we complete hash1 with
padding and calculate hash2 (the final tag).

Signed-off-by: Gowtham Suresh Kumar <gowtham.sureshkumar@arm.com>